+++ package.json
@@ -4,7 +4,7 @@
   "version": "0.221.0",
   "dependencies": {
     "@vtex/gatsby-plugin-graphql": "^0.373.16",
-    "@vtex/gatsby-plugin-nginx": "https://pkg.csb.dev/vtex/faststore/commit/7f478c22/@vtex/gatsby-plugin-nginx",
+    "@vtex/gatsby-plugin-nginx": "^0.373.16",
     "@vtex/gatsby-plugin-thumbor": "0.373.16",
     "@vtex/gatsby-source-vtex": "^0.373.16",
     "@vtex/store-sdk": "https://pkg.csb.dev/vtex/faststore/commit/add8ed84/@vtex/store-sdk",
